Add a 'sudo warmup' to command scripts.
I'm running into a problem where script execution is sort-of-kind-of pausing on the first 'sudo' command; while sudo waits for a password, some or all of the remaining script gets read into stdin, making it impossible to enter a correct password and also causing script execution to be inconsistent.

If the password is entered correctly on the second attempt, some of the script commands don't get re-executed. It's a mess.

I hope that using 'sudo' in an if...then block will resolve this.