Allow .md files to be imported directly as scripts.
I didn't do this initially because I imagined plain markdown files might be imported as viewable documentation someday. But, that hasn't happened yet, and if it does then "import document" is available, and in the meantime I've moved my executable markdown stuff to a new system and I'd like to be able to invoke Golem against it directly.