Add source script hash verification to cache check.
This allows files in the cache to have different timestamps from their source scripts as long as the source script's content hasn't changed. This should make it easier to copy a golem to remotes (to run a remote golem against a remote remote) without incurring a script re-import on systems that might not have mdsh installed.